requirements.txt
numpy
matplotlib
pyyaml
fastapi
uvicorn
